New hook for searchOrg, new component for affiliation modal
